fix component Like posts usere
